+++ astro.config.ts
@@ -12,7 +12,9 @@ export default defineConfig({
 				provider: fontProviders.google(),
 			},
 		],
-		responsiveImages: true,
+	},
+	image: {
+		responsiveStyles: true,
 	},
 	integrations: [
 		konamiEmojiBlast(),
@@ -21,6 +23,7 @@ export default defineConfig({
 			components: {
 				Footer: "./src/components/Footer.astro",
 				Head: "./src/components/Head.astro",
+				SiteTitle: "./src/components/SiteTitle.astro",
 				TwoColumnContent: "./src/components/TwoColumnContent.astro",
 			},
 			customCss: ["./src/styles/global.css"],